Make parameter values optional and move role tests to write service (red)
