+++ Core.ApplicationServices/Model/Organizations/Write/MasterDataRoles/DataResponsibleUpdateParameters.cs
@@ -1,10 +1,11 @@
-﻿using Core.ApplicationServices.Model.Shared;
+﻿using Core.Abstractions.Types;
+using Core.ApplicationServices.Model.Shared;
 
 namespace Core.ApplicationServices.Model.Organizations.Write.MasterDataRoles;
 
 public class DataResponsibleUpdateParameters: OrganizationMasterDataRoleUpdateParameters
 {
-    public OptionalValueChange<string> Cvr { get; set; }
-    public OptionalValueChange<string> Phone { get; set; }
-    public OptionalValueChange<string> Address { get; set; }
+    public OptionalValueChange<Maybe<string>> Cvr { get; set; }
+    public OptionalValueChange<Maybe<string>> Phone { get; set; }
+    public OptionalValueChange<Maybe<string>> Address { get; set; }
 }
