+++ Core.ApplicationServices/Model/Organizations/Write/MasterDataRoles/OrganizationMasterDataRoleUpdateParameters.cs
@@ -1,10 +1,11 @@
-﻿using Core.ApplicationServices.Model.Shared;
+﻿using Core.Abstractions.Types;
+using Core.ApplicationServices.Model.Shared;
 
 namespace Core.ApplicationServices.Model.Organizations.Write.MasterDataRoles
 {
     public class OrganizationMasterDataRoleUpdateParameters
     {
-        public OptionalValueChange<string> Name { get; set; }
-        public OptionalValueChange<string> Email { get; set; }
+        public OptionalValueChange<Maybe<string>> Name { get; set; }
+        public OptionalValueChange<Maybe<string>> Email { get; set; }
     }
 }
