+++ Core.ApplicationServices/Organizations/Write/IOrganizationWriteService.cs
@@ -1,5 +1,7 @@
 ﻿using Core.Abstractions.Types;
+using Core.ApplicationServices.Model.Organizations;
 using Core.ApplicationServices.Model.Organizations.Write;
+using Core.ApplicationServices.Model.Organizations.Write.MasterDataRoles;
 using Core.DomainModel.Organization;
 using System;
 
@@ -9,5 +11,8 @@ namespace Core.ApplicationServices.Organizations.Write
     {
         Result<Organization, OperationError> UpdateMasterData(Guid organizationUuid,
             OrganizationMasterDataUpdateParameters parameters);
+
+        Result<OrganizationMasterDataRoles, OperationError> UpsertOrganizationMasterDataRoles(Guid organizationUuid, OrganizationMasterDataRolesUpdateParameters updateParameters);
+
     }
 }
