+++ Tests.Unit.Presentation.Web/Services/OrganizationServiceTest.cs
@@ -6,9 +6,6 @@ using Core.Abstractions.Extensions;
 using Core.Abstractions.Types;
 using Core.ApplicationServices.Authorization;
 using Core.ApplicationServices.Authorization.Permissions;
-using Core.ApplicationServices.Model.Organizations.Write;
-using Core.ApplicationServices.Model.Organizations.Write.MasterDataRoles;
-using Core.ApplicationServices.Model.Shared;
 using Core.ApplicationServices.Organizations;
 using Core.DomainModel;
 using Core.DomainModel.Events;
@@ -880,263 +877,6 @@ namespace Tests.Unit.Presentation.Web.Services
             Assert.True(result.Failed);
             var error = result.Error;
             Assert.Equal(OperationFailure.BadInput, error.FailureType);
-        }
-
-        [Fact]
-        public void Update_Master_Data_Roles_Returns_Bad_Input_If_Invalid_Uuid()
-        {
-            var invalidOrganizationUuid = A<Guid>();
-            _identityResolver.Setup(_ =>
-                    _.ResolveDbId<Organization>(invalidOrganizationUuid))
-                .Returns(Maybe<int>.None);
-
-            var result =
-                _sut.UpsertOrganizationMasterDataRoles(invalidOrganizationUuid, new OrganizationMasterDataRolesUpdateParameters());
-
-            Assert.True(result.Failed);
-            var error = result.Error;
-            Assert.Equal(OperationFailure.BadInput, error.FailureType);
-        }
-
-        [Fact]
-        public void Update_Master_Data_Roles_Returns_Forbidden_If_Unauthorized_To_Modify_Data_Responsible()
-        {
-            var org = CreateOrganization();
-            var orgId = org.Id;
-            var updateParameters = SetupUpdateMasterDataRoles(orgId);
-            _identityResolver.Setup(_ =>
-                    _.ResolveDbId<Organization>(org.Uuid))
-                .Returns(orgId);
-            _authorizationContext.Setup(_ =>
-                    _.AllowModify(It.IsAny<ContactPerson>()))
-                .Returns(true);
-            _authorizationContext.Setup(_ =>
-                    _.AllowModify(It.IsAny<DataResponsible>()))
-                .Returns(false);
-            _authorizationContext.Setup(_ =>
-                    _.AllowModify(It.IsAny<DataProtectionAdvisor>()))
-                .Returns(true);
-
-            var transaction = new Mock<IDatabaseTransaction>();
-            _transactionManager.Setup(x => x.Begin()).Returns(transaction.Object);
-
-            var result =
-                _sut.UpsertOrganizationMasterDataRoles(org.Uuid, updateParameters);
-
-            Assert.True(result.Failed);
-            var error = result.Error;
-            Assert.Equal(OperationFailure.Forbidden, error.FailureType);
-        }
-
-        [Fact]
-        public void Update_Master_Data_Roles_Returns_Forbidden_If_Unauthorized_To_Modify_Contact_Person()
-        {
-            var org = CreateOrganization();
-            var orgId = org.Id;
-            var updateParameters = SetupUpdateMasterDataRoles(orgId);
-            _identityResolver.Setup(_ =>
-                    _.ResolveDbId<Organization>(org.Uuid))
-                .Returns(orgId);
-            _authorizationContext.Setup(_ =>
-                    _.AllowModify(It.IsAny<ContactPerson>()))
-                .Returns(false);
-            _authorizationContext.Setup(_ =>
-                    _.AllowModify(It.IsAny<DataResponsible>()))
-                .Returns(true);
-            _authorizationContext.Setup(_ =>
-                    _.AllowModify(It.IsAny<DataProtectionAdvisor>()))
-                .Returns(true);
-
-            var transaction = new Mock<IDatabaseTransaction>();
-            _transactionManager.Setup(x => x.Begin()).Returns(transaction.Object);
-
-            var result =
-                _sut.UpsertOrganizationMasterDataRoles(org.Uuid, updateParameters);
-
-            Assert.True(result.Failed);
-            var error = result.Error;
-            Assert.Equal(OperationFailure.Forbidden, error.FailureType);
-        }
-
-        [Fact]
-        public void Update_Master_Data_Roles_Returns_Forbidden_If_Unauthorized_To_Modify_Data_Protection_Advisor()
-        {
-            var org = CreateOrganization();
-            var orgId = org.Id;
-            var updateParameters = SetupUpdateMasterDataRoles(orgId);
-            _identityResolver.Setup(_ =>
-                    _.ResolveDbId<Organization>(org.Uuid))
-                .Returns(orgId);
-            _authorizationContext.Setup(_ =>
-                    _.AllowModify(It.IsAny<ContactPerson>()))
-                .Returns(true);
-            _authorizationContext.Setup(_ =>
-                    _.AllowModify(It.IsAny<DataResponsible>()))
-                .Returns(true);
-            _authorizationContext.Setup(_ =>
-                    _.AllowModify(It.IsAny<DataProtectionAdvisor>()))
-                .Returns(false);
-
-            var transaction = new Mock<IDatabaseTransaction>();
-            _transactionManager.Setup(x => x.Begin()).Returns(transaction.Object);
-
-            var result =
-                _sut.UpsertOrganizationMasterDataRoles(org.Uuid, updateParameters);
-
-            Assert.True(result.Failed);
-            var error = result.Error;
-            Assert.Equal(OperationFailure.Forbidden, error.FailureType);
-        }
-
-        [Fact]
-        public void Can_Update_Master_Data_Roles()
-        {
-            var org = CreateOrganization();
-            var orgId = org.Id;
-            var expectedContactPerson = SetupGetMasterDataRolesContactPerson(orgId);
-            var expectedDataResponsible = SetupGetMasterDataRolesDataResponsible(orgId);
-            var expectedDataProtectionAdvisor = SetupGetMasterDataRolesDataProtectionAdvisor(orgId);
-            var updateParameters = SetupUpdateMasterDataRoles(orgId, expectedContactPerson, expectedDataResponsible, expectedDataProtectionAdvisor);
-            _identityResolver.Setup(_ =>
-                    _.ResolveDbId<Organization>(org.Uuid))
-                .Returns(expectedContactPerson.OrganizationId);
-
-            _authorizationContext.Setup(_ =>
-                    _.AllowModify(It.IsAny<ContactPerson>()))
-                .Returns(true);
-            _authorizationContext.Setup(_ =>
-                    _.AllowModify(It.IsAny<DataResponsible>()))
-                .Returns(true);
-            _authorizationContext.Setup(_ =>
-                    _.AllowModify(It.IsAny<DataProtectionAdvisor>()))
-                .Returns(true);
-
-            var transaction = new Mock<IDatabaseTransaction>();
-            _transactionManager.Setup(x => x.Begin()).Returns(transaction.Object);
-
-            var result = _sut.UpsertOrganizationMasterDataRoles(org.Uuid, updateParameters);
-
-            Assert.True(result.Ok);
-            var value = result.Value;
-            AssertContactPerson(expectedContactPerson, value.ContactPerson);
-            AssertDataResponsible(expectedDataResponsible, value.DataResponsible);
-            AssertDataProtectionAdvisor(expectedDataProtectionAdvisor, value.DataProtectionAdvisor);
-        }
-
-        [Fact]
-        public void Update_Creates_Master_Data_Roles_If_Not_Found()
-        {
-            var org = CreateOrganization();
-            var orgId = org.Id;
-            var expectedContactPerson = SetupGetMasterDataRolesContactPerson(orgId);
-            var expectedDataResponsible = SetupGetMasterDataRolesDataResponsible(orgId);
-            var expectedDataProtectionAdvisor = SetupGetMasterDataRolesDataProtectionAdvisor(orgId);
-            var updateParameters = new OrganizationMasterDataRolesUpdateParameters
-            {
-                ContactPerson = new ContactPersonUpdateParameters()
-                {
-                    Email = OptionalValueChange<string>.With(expectedContactPerson.Email),
-                    Name = OptionalValueChange<string>.With(expectedContactPerson.Name),
-                    LastName = OptionalValueChange<string>.With(expectedContactPerson.LastName),
-                    PhoneNumber = OptionalValueChange<string>.With(expectedContactPerson.PhoneNumber),
-                },
-                DataResponsible = new DataResponsibleUpdateParameters()
-                {
-                    Email = OptionalValueChange<string>.With(expectedDataResponsible.Email),
-                    Name = OptionalValueChange<string>.With(expectedDataResponsible.Name),
-                    Cvr = OptionalValueChange<string>.With(expectedDataResponsible.Cvr),
-                    Address = OptionalValueChange<string>.With(expectedDataResponsible.Adress),
-                    Phone = OptionalValueChange<string>.With(expectedDataResponsible.Phone)
-                },
-                DataProtectionAdvisor = new DataProtectionAdvisorUpdateParameters()
-                {
-                    Email = OptionalValueChange<string>.With(expectedDataProtectionAdvisor.Email),
-                    Name = OptionalValueChange<string>.With(expectedDataProtectionAdvisor.Name),
-                    Cvr = OptionalValueChange<string>.With(expectedDataProtectionAdvisor.Cvr),
-                    Address = OptionalValueChange<string>.With(expectedDataProtectionAdvisor.Adress),
-                    Phone = OptionalValueChange<string>.With(expectedDataProtectionAdvisor.Phone)
-                }
-            }; 
-            _identityResolver.Setup(_ =>
-                    _.ResolveDbId<Organization>(org.Uuid))
-                .Returns(expectedContactPerson.OrganizationId);
-
-            _authorizationContext.Setup(_ =>
-                    _.AllowModify(It.IsAny<ContactPerson>()))
-                .Returns(true);
-            _authorizationContext.Setup(_ =>
-                    _.AllowModify(It.IsAny<DataResponsible>()))
-                .Returns(true);
-            _authorizationContext.Setup(_ =>
-                    _.AllowModify(It.IsAny<DataProtectionAdvisor>()))
-                .Returns(true);
-
-            _contactPersonRepository.Setup(_ =>
-                    _.AsQueryable())
-                .Returns(new List<ContactPerson>().AsQueryable());
-            _dataResponsibleRepository.Setup(_ =>
-                    _.AsQueryable())
-                .Returns(new List<DataResponsible>().AsQueryable());
-            _dataProtectionAdvisorRepository.Setup(_ =>
-                    _.AsQueryable())
-                .Returns(new List<DataProtectionAdvisor>().AsQueryable());
-
-            var transaction = new Mock<IDatabaseTransaction>();
-            _transactionManager.Setup(x => x.Begin()).Returns(transaction.Object);
-
-            var result = _sut.UpsertOrganizationMasterDataRoles(org.Uuid, updateParameters);
-
-            Assert.True(result.Ok);
-            _contactPersonRepository.Verify(_ => _.Insert(It.IsAny<ContactPerson>()));
-            _dataResponsibleRepository.Verify(_ => _.Insert(It.IsAny<DataResponsible>()));
-            _dataProtectionAdvisorRepository.Verify(_ => _.Insert(It.IsAny<DataProtectionAdvisor>()));
-            var value = result.Value;
-            AssertContactPerson(expectedContactPerson, value.ContactPerson);
-            AssertDataResponsible(expectedDataResponsible, value.DataResponsible);
-            AssertDataProtectionAdvisor(expectedDataProtectionAdvisor, value.DataProtectionAdvisor);
-        }
-
-        private OrganizationMasterDataRolesUpdateParameters SetupUpdateMasterDataRoles(int orgId,
-            ContactPerson cp = null, DataResponsible dr = null, DataProtectionAdvisor dpa = null)
-        {
-            var expectedContactPerson = cp ?? SetupGetMasterDataRolesContactPerson(orgId);
-            var expectedDataResponsible = dr ?? SetupGetMasterDataRolesDataResponsible(orgId);
-            var expectedDataProtectionAdvisor = dpa ?? SetupGetMasterDataRolesDataProtectionAdvisor(orgId);
-            _contactPersonRepository.Setup(_ =>
-                    _.AsQueryable())
-                .Returns(new List<ContactPerson> { expectedContactPerson }.AsQueryable());
-            _dataResponsibleRepository.Setup(_ =>
-                    _.AsQueryable())
-                .Returns(new List<DataResponsible> { expectedDataResponsible }.AsQueryable());
-            _dataProtectionAdvisorRepository.Setup(_ =>
-                    _.AsQueryable())
-                .Returns(new List<DataProtectionAdvisor> { expectedDataProtectionAdvisor }.AsQueryable());
-            return new OrganizationMasterDataRolesUpdateParameters
-            {
-                ContactPerson = new ContactPersonUpdateParameters()
-                {
-                    Email = OptionalValueChange<string>.With(expectedContactPerson.Email),
-                    Name = OptionalValueChange<string>.With(expectedContactPerson.Name),
-                    LastName = OptionalValueChange<string>.With(expectedContactPerson.LastName),
-                    PhoneNumber = OptionalValueChange<string>.With(expectedContactPerson.PhoneNumber),
-                },
-                DataResponsible = new DataResponsibleUpdateParameters()
-                {
-                    Email = OptionalValueChange<string>.With(expectedContactPerson.Email),
-                    Name = OptionalValueChange<string>.With(expectedContactPerson.Name),
-                    Cvr = OptionalValueChange<string>.With(expectedDataResponsible.Cvr),
-                    Address = OptionalValueChange<string>.With(expectedDataResponsible.Adress),
-                    Phone = OptionalValueChange<string>.With(expectedDataResponsible.Phone)
-                },
-                DataProtectionAdvisor = new DataProtectionAdvisorUpdateParameters()
-                {
-                    Email = OptionalValueChange<string>.With(expectedContactPerson.Email),
-                    Name = OptionalValueChange<string>.With(expectedContactPerson.Name),
-                    Cvr = OptionalValueChange<string>.With(expectedDataResponsible.Cvr),
-                    Address = OptionalValueChange<string>.With(expectedDataResponsible.Adress),
-                    Phone = OptionalValueChange<string>.With(expectedDataResponsible.Phone)
-                }
-            };
         }
 
         private void AssertContactPerson(ContactPerson expected, ContactPerson actual)
