+++ Tests.Unit.Presentation.Web/Services/OrganizationWriteServiceTest.cs
@@ -13,6 +13,12 @@ using Core.ApplicationServices.Organizations.Write;
 using Tests.Toolkit.Patterns;
 using Xunit;
 using Core.DomainModel.Events;
+using Core.DomainServices.Generic;
+using Core.ApplicationServices.Model.Organizations.Write.MasterDataRoles;
+using Core.DomainModel;
+using System.Collections.Generic;
+using System.Linq;
+using Core.DomainServices;
 
 namespace Tests.Unit.Presentation.Web.Services
 {
@@ -23,6 +29,10 @@ namespace Tests.Unit.Presentation.Web.Services
         private readonly Mock<IDomainEvents> _domainEvents;
         private readonly Mock<IOrganizationRepository> _organizationrepository;
         private readonly Mock<IOrganizationService> _organizationService;
+        private readonly Mock<IEntityIdentityResolver> _identityResolver;
+        private readonly Mock<IGenericRepository<ContactPerson>> _contactPersonRepository;
+        private readonly Mock<IGenericRepository<DataResponsible>> _dataResponsibleRepository;
+        private readonly Mock<IGenericRepository<DataProtectionAdvisor>> _dataProtectionAdvisorRepository;
         private readonly OrganizationWriteService _sut;
 
 
@@ -33,11 +43,19 @@ namespace Tests.Unit.Presentation.Web.Services
             _domainEvents = new Mock<IDomainEvents>();  
             _organizationrepository = new Mock<IOrganizationRepository>();
             _organizationService = new Mock<IOrganizationService>();
+            _identityResolver = new Mock<IEntityIdentityResolver>();
+            _contactPersonRepository = new Mock<IGenericRepository<ContactPerson>>();
+            _dataResponsibleRepository = new Mock<IGenericRepository<DataResponsible>>();
+            _dataProtectionAdvisorRepository = new Mock<IGenericRepository<DataProtectionAdvisor>>();
             _sut = new OrganizationWriteService(_transactionManager.Object,
                 _domainEvents.Object,
                 _organizationService.Object,
                 _authorizationContext.Object,
-                _organizationrepository.Object);
+                _organizationrepository.Object,
+                _identityResolver.Object,
+                _contactPersonRepository.Object,
+                _dataResponsibleRepository.Object,
+                _dataProtectionAdvisorRepository.Object);
         }
 
         [Fact]
@@ -140,6 +158,331 @@ namespace Tests.Unit.Presentation.Web.Services
             Assert.Null(updatedOrganization.Adress);
             Assert.Null(updatedOrganization.Email);
             _organizationrepository.Verify(_ => _.Update(organization.Object));
+        }
+
+        [Fact]
+        public void Update_Master_Data_Roles_Returns_Bad_Input_If_Invalid_Uuid()
+        {
+            var invalidOrganizationUuid = A<Guid>();
+            _identityResolver.Setup(_ =>
+                    _.ResolveDbId<Organization>(invalidOrganizationUuid))
+                .Returns(Maybe<int>.None);
+
+            var result =
+                _sut.UpsertOrganizationMasterDataRoles(invalidOrganizationUuid, new OrganizationMasterDataRolesUpdateParameters());
+
+            Assert.True(result.Failed);
+            var error = result.Error;
+            Assert.Equal(OperationFailure.BadInput, error.FailureType);
+        }
+
+        [Fact]
+        public void Update_Master_Data_Roles_Returns_Forbidden_If_Unauthorized_To_Modify_Data_Responsible()
+        {
+            var org = CreateOrganization();
+            var orgId = org.Id;
+            var updateParameters = SetupUpdateMasterDataRoles(orgId);
+            _identityResolver.Setup(_ =>
+                    _.ResolveDbId<Organization>(org.Uuid))
+                .Returns(orgId);
+            _authorizationContext.Setup(_ =>
+                    _.AllowModify(It.IsAny<ContactPerson>()))
+                .Returns(true);
+            _authorizationContext.Setup(_ =>
+                    _.AllowModify(It.IsAny<DataResponsible>()))
+                .Returns(false);
+            _authorizationContext.Setup(_ =>
+                    _.AllowModify(It.IsAny<DataProtectionAdvisor>()))
+                .Returns(true);
+
+            var transaction = new Mock<IDatabaseTransaction>();
+            _transactionManager.Setup(x => x.Begin()).Returns(transaction.Object);
+
+            var result =
+                _sut.UpsertOrganizationMasterDataRoles(org.Uuid, updateParameters);
+
+            Assert.True(result.Failed);
+            var error = result.Error;
+            Assert.Equal(OperationFailure.Forbidden, error.FailureType);
+        }
+
+        [Fact]
+        public void Update_Master_Data_Roles_Returns_Forbidden_If_Unauthorized_To_Modify_Contact_Person()
+        {
+            var org = CreateOrganization();
+            var orgId = org.Id;
+            var updateParameters = SetupUpdateMasterDataRoles(orgId);
+            _identityResolver.Setup(_ =>
+                    _.ResolveDbId<Organization>(org.Uuid))
+                .Returns(orgId);
+            _authorizationContext.Setup(_ =>
+                    _.AllowModify(It.IsAny<ContactPerson>()))
+                .Returns(false);
+            _authorizationContext.Setup(_ =>
+                    _.AllowModify(It.IsAny<DataResponsible>()))
+                .Returns(true);
+            _authorizationContext.Setup(_ =>
+                    _.AllowModify(It.IsAny<DataProtectionAdvisor>()))
+                .Returns(true);
+
+            var transaction = new Mock<IDatabaseTransaction>();
+            _transactionManager.Setup(x => x.Begin()).Returns(transaction.Object);
+
+            var result =
+                _sut.UpsertOrganizationMasterDataRoles(org.Uuid, updateParameters);
+
+            Assert.True(result.Failed);
+            var error = result.Error;
+            Assert.Equal(OperationFailure.Forbidden, error.FailureType);
+        }
+
+        [Fact]
+        public void Update_Master_Data_Roles_Returns_Forbidden_If_Unauthorized_To_Modify_Data_Protection_Advisor()
+        {
+            var org = CreateOrganization();
+            var orgId = org.Id;
+            var updateParameters = SetupUpdateMasterDataRoles(orgId);
+            _identityResolver.Setup(_ =>
+                    _.ResolveDbId<Organization>(org.Uuid))
+                .Returns(orgId);
+            _authorizationContext.Setup(_ =>
+                    _.AllowModify(It.IsAny<ContactPerson>()))
+                .Returns(true);
+            _authorizationContext.Setup(_ =>
+                    _.AllowModify(It.IsAny<DataResponsible>()))
+                .Returns(true);
+            _authorizationContext.Setup(_ =>
+                    _.AllowModify(It.IsAny<DataProtectionAdvisor>()))
+                .Returns(false);
+
+            var transaction = new Mock<IDatabaseTransaction>();
+            _transactionManager.Setup(x => x.Begin()).Returns(transaction.Object);
+
+            var result =
+                _sut.UpsertOrganizationMasterDataRoles(org.Uuid, updateParameters);
+
+            Assert.True(result.Failed);
+            var error = result.Error;
+            Assert.Equal(OperationFailure.Forbidden, error.FailureType);
+        }
+
+        [Fact]
+        public void Can_Update_Master_Data_Roles()
+        {
+            var org = CreateOrganization();
+            var orgId = org.Id;
+            var expectedContactPerson = SetupGetMasterDataRolesContactPerson(orgId);
+            var expectedDataResponsible = SetupGetMasterDataRolesDataResponsible(orgId);
+            var expectedDataProtectionAdvisor = SetupGetMasterDataRolesDataProtectionAdvisor(orgId);
+            var updateParameters = SetupUpdateMasterDataRoles(orgId, expectedContactPerson, expectedDataResponsible, expectedDataProtectionAdvisor);
+            _identityResolver.Setup(_ =>
+                    _.ResolveDbId<Organization>(org.Uuid))
+                .Returns(expectedContactPerson.OrganizationId);
+
+            _authorizationContext.Setup(_ =>
+                    _.AllowModify(It.IsAny<ContactPerson>()))
+                .Returns(true);
+            _authorizationContext.Setup(_ =>
+                    _.AllowModify(It.IsAny<DataResponsible>()))
+                .Returns(true);
+            _authorizationContext.Setup(_ =>
+                    _.AllowModify(It.IsAny<DataProtectionAdvisor>()))
+                .Returns(true);
+
+            var transaction = new Mock<IDatabaseTransaction>();
+            _transactionManager.Setup(x => x.Begin()).Returns(transaction.Object);
+
+            var result = _sut.UpsertOrganizationMasterDataRoles(org.Uuid, updateParameters);
+
+            Assert.True(result.Ok);
+            var value = result.Value;
+            AssertContactPerson(expectedContactPerson, value.ContactPerson);
+            AssertDataResponsible(expectedDataResponsible, value.DataResponsible);
+            AssertDataProtectionAdvisor(expectedDataProtectionAdvisor, value.DataProtectionAdvisor);
+        }
+
+        [Fact]
+        public void Update_Creates_Master_Data_Roles_If_Not_Found()
+        {
+            var org = CreateOrganization();
+            var orgId = org.Id;
+            var expectedContactPerson = SetupGetMasterDataRolesContactPerson(orgId);
+            var expectedDataResponsible = SetupGetMasterDataRolesDataResponsible(orgId);
+            var expectedDataProtectionAdvisor = SetupGetMasterDataRolesDataProtectionAdvisor(orgId);
+            var updateParameters = GetRolesUpdateParameters(expectedContactPerson, expectedDataResponsible,
+                expectedDataProtectionAdvisor);
+
+            _identityResolver.Setup(_ =>
+                    _.ResolveDbId<Organization>(org.Uuid))
+                .Returns(expectedContactPerson.OrganizationId);
+
+            _authorizationContext.Setup(_ =>
+                    _.AllowModify(It.IsAny<ContactPerson>()))
+                .Returns(true);
+            _authorizationContext.Setup(_ =>
+                    _.AllowModify(It.IsAny<DataResponsible>()))
+                .Returns(true);
+            _authorizationContext.Setup(_ =>
+                    _.AllowModify(It.IsAny<DataProtectionAdvisor>()))
+                .Returns(true);
+
+            _contactPersonRepository.Setup(_ =>
+                    _.AsQueryable())
+                .Returns(new List<ContactPerson>().AsQueryable());
+            _dataResponsibleRepository.Setup(_ =>
+                    _.AsQueryable())
+                .Returns(new List<DataResponsible>().AsQueryable());
+            _dataProtectionAdvisorRepository.Setup(_ =>
+                    _.AsQueryable())
+                .Returns(new List<DataProtectionAdvisor>().AsQueryable());
+
+            var transaction = new Mock<IDatabaseTransaction>();
+            _transactionManager.Setup(x => x.Begin()).Returns(transaction.Object);
+
+            var result = _sut.UpsertOrganizationMasterDataRoles(org.Uuid, updateParameters);
+
+            Assert.True(result.Ok);
+            _contactPersonRepository.Verify(_ => _.Insert(It.IsAny<ContactPerson>()));
+            _dataResponsibleRepository.Verify(_ => _.Insert(It.IsAny<DataResponsible>()));
+            _dataProtectionAdvisorRepository.Verify(_ => _.Insert(It.IsAny<DataProtectionAdvisor>()));
+            var value = result.Value;
+            AssertContactPerson(expectedContactPerson, value.ContactPerson);
+            AssertDataResponsible(expectedDataResponsible, value.DataResponsible);
+            AssertDataProtectionAdvisor(expectedDataProtectionAdvisor, value.DataProtectionAdvisor);
+        }
+
+        private OrganizationMasterDataRolesUpdateParameters SetupUpdateMasterDataRoles(int orgId,
+            ContactPerson cp = null, DataResponsible dr = null, DataProtectionAdvisor dpa = null)
+        {
+            var expectedContactPerson = cp ?? SetupGetMasterDataRolesContactPerson(orgId);
+            var expectedDataResponsible = dr ?? SetupGetMasterDataRolesDataResponsible(orgId);
+            var expectedDataProtectionAdvisor = dpa ?? SetupGetMasterDataRolesDataProtectionAdvisor(orgId);
+            _contactPersonRepository.Setup(_ =>
+                    _.AsQueryable())
+                .Returns(new List<ContactPerson> { expectedContactPerson }.AsQueryable());
+            _dataResponsibleRepository.Setup(_ =>
+                    _.AsQueryable())
+                .Returns(new List<DataResponsible> { expectedDataResponsible }.AsQueryable());
+            _dataProtectionAdvisorRepository.Setup(_ =>
+                    _.AsQueryable())
+                .Returns(new List<DataProtectionAdvisor> { expectedDataProtectionAdvisor }.AsQueryable());
+            return GetRolesUpdateParameters(expectedContactPerson, expectedDataResponsible,
+                expectedDataProtectionAdvisor);
+        }
+
+        private OrganizationMasterDataRolesUpdateParameters GetRolesUpdateParameters(ContactPerson expectedContactPerson,
+            DataResponsible expectedDataResponsible, DataProtectionAdvisor expectedDataProtectionAdvisor)
+        {
+            return new OrganizationMasterDataRolesUpdateParameters
+            {
+                ContactPerson = new ContactPersonUpdateParameters()
+                {
+                    Email = OptionalValueChange<Maybe<string>>.With(Maybe<string>.Some(expectedContactPerson.Email)),
+                    Name = OptionalValueChange<Maybe<string>>.With(Maybe<string>.Some(expectedContactPerson.Name)),
+                    LastName = OptionalValueChange<Maybe<string>>.With(Maybe<string>.Some(expectedContactPerson.LastName)),
+                    PhoneNumber = OptionalValueChange<Maybe<string>>.With(Maybe<string>.Some(expectedContactPerson.PhoneNumber)),
+                },
+                DataResponsible = new DataResponsibleUpdateParameters()
+                {
+                    Email = OptionalValueChange<Maybe<string>>.With(Maybe<string>.Some(expectedDataResponsible.Email)),
+                    Name = OptionalValueChange<Maybe<string>>.With(Maybe<string>.Some(expectedDataResponsible.Name)),
+                    Cvr = OptionalValueChange<Maybe<string>>.With(Maybe<string>.Some(expectedDataResponsible.Cvr)),
+                    Address = OptionalValueChange<Maybe<string>>.With(Maybe<string>.Some(expectedDataResponsible.Adress)),
+                    Phone = OptionalValueChange<Maybe<string>>.With(Maybe<string>.Some(expectedDataResponsible.Phone))
+                },
+                DataProtectionAdvisor = new DataProtectionAdvisorUpdateParameters()
+                {
+                    Email = OptionalValueChange<Maybe<string>>.With(Maybe<string>.Some(expectedDataProtectionAdvisor.Email)),
+                    Name = OptionalValueChange<Maybe<string>>.With(Maybe<string>.Some(expectedDataProtectionAdvisor.Name)),
+                    Cvr = OptionalValueChange<Maybe<string>>.With(Maybe<string>.Some(expectedDataProtectionAdvisor.Cvr)),
+                    Address = OptionalValueChange<Maybe<string>>.With(Maybe<string>.Some(expectedDataProtectionAdvisor.Adress)),
+                    Phone = OptionalValueChange<Maybe<string>>.With(Maybe<string>.Some(expectedDataProtectionAdvisor.Phone))
+                }
+            };
+        }
+        private Organization CreateOrganization()
+        {
+            var organizationId = A<Guid>();
+            var organization = new Organization() { Uuid = organizationId, Id = A<int>() };
+            return organization;
+        }
+
+        private DataProtectionAdvisor SetupGetMasterDataRolesDataProtectionAdvisor(int orgId)
+        {
+            var expectedDataProtectionAdvisor = new DataProtectionAdvisor()
+            {
+                Email = A<string>(),
+                Name = A<string>(),
+                Cvr = A<string>(),
+                Adress = A<string>(),
+                OrganizationId = orgId,
+                Phone = A<string>(),
+                Id = A<int>(),
+            };
+            _dataProtectionAdvisorRepository.Setup(_ =>
+                    _.AsQueryable())
+                .Returns(new List<DataProtectionAdvisor> { expectedDataProtectionAdvisor }.AsQueryable());
+            return expectedDataProtectionAdvisor;
+        }
+
+        private ContactPerson SetupGetMasterDataRolesContactPerson(int orgId)
+        {
+            var expectedContactPerson = new ContactPerson
+            {
+                Email = A<string>(),
+                Name = A<string>(),
+                PhoneNumber = A<string>(),
+                OrganizationId = orgId,
+                Id = A<int>(),
+            };
+            _contactPersonRepository.Setup(_ =>
+                    _.AsQueryable())
+                .Returns(new List<ContactPerson> { expectedContactPerson }.AsQueryable());
+            return expectedContactPerson;
+        }
+
+        private DataResponsible SetupGetMasterDataRolesDataResponsible(int orgId)
+        {
+            var expectedDataResponsible = new DataResponsible
+            {
+                Email = A<string>(),
+                Name = A<string>(),
+                Cvr = A<string>(),
+                Adress = A<string>(),
+                Phone = A<string>(),
+                OrganizationId = orgId,
+                Id = A<int>(),
+            };
+            _dataResponsibleRepository.Setup(_ =>
+                    _.AsQueryable())
+                .Returns(new List<DataResponsible> { expectedDataResponsible }.AsQueryable());
+            return expectedDataResponsible;
+        }
+
+        private void AssertContactPerson(ContactPerson expected, ContactPerson actual)
+        {
+            Assert.Equal(expected.Email, actual.Email);
+            Assert.Equal(expected.Name, actual.Name);
+            Assert.Equal(expected.PhoneNumber, actual.PhoneNumber);
+            Assert.Equal(expected.OrganizationId, actual.OrganizationId);
+        }
+
+        private void AssertDataResponsible(DataResponsible expected, DataResponsible actual)
+        {
+            Assert.Equal(expected.Email, actual.Email);
+            Assert.Equal(expected.Name, actual.Name);
+            Assert.Equal(expected.Cvr, actual.Cvr);
+            Assert.Equal(expected.OrganizationId, actual.OrganizationId);
+            Assert.Equal(expected.Adress, actual.Adress);
+        }
+
+        private void AssertDataProtectionAdvisor(DataProtectionAdvisor expected, DataProtectionAdvisor actual)
+        {
+            Assert.Equal(expected.Email, actual.Email);
+            Assert.Equal(expected.Name, actual.Name);
+            Assert.Equal(expected.Cvr, actual.Cvr);
+            Assert.Equal(expected.OrganizationId, actual.OrganizationId);
+            Assert.Equal(expected.Adress, actual.Adress);
         }
     }
 }
